CA-274957: simplify and reorganise Nvidia VGPU compatibility check
The function `assert_destination_pgpu_is_compatible_with_vm`
had a heuristic to select the destination pgpu that was
filtering over vgpu implementation, however this is `nvidia`
for any Nvidia card, resulting in potentially meaningless
choices.

The function was also assuming that the user-defined
vgpu_map was complete (all vgpus would be mapped) but
this is often not the case.

The function was called in two migration places, in one
we had preselected the pgpu, making any heuristic
a potential source of errors.

This commit splits the check into two assertions:
- `assert_destination_pgpu_is_compatible_with_vm`, to be called
  when the destination pgpu has already been selected
- `assert_destination_has_pgpu_compatible_with_vm`, using
  a heuristic (based on the one used by vm-import) to
  select a suitable pgpu group.

The new heuristic selects a gpu group by filtering
over the gpu types, making sure that the destination
group is compatible with the origin group.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>