CMP: add decompression to Domain.restore
Decompress the stream. This is currently unconditional but needs to be
made conditional based on the header of the suspend stream. It contains
a (new) field that signals if compression is used. It is absent on older
streams and defaults to None.

We need to be careful: the header is uncompressed and needs to be read
first. After that the stream is compressed and we need to read it using
the decompress filter, which is a gzip process connected via a pipe.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>