CA-379472 log more block_device_io messages to info
The debug log level is disabled by default. Since we are strugglign to
understand why this process is slow to startup or other reasons why xapi
fails to connect to its socket, let's elevate more logging to the info
level. We want to avoid logging the regular case where data is
transferred but are interested in the setup and teardown case.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>